Made necessary modification to recompile with boost_1_56 and gcc 8.21. changed time.h to ctime and math to cmath. There were problems with getcwd but by including boost file boost/filesystem.hpp fixed this.
Then offcourse the necessary settings to eclipse to include the boost_system and boost_programoptions library, gsl, m cblas

+++ simAllocation-cAMP/InputVectorHandling.cpp
@@ -375,7 +375,7 @@ void makeGN2TestVectorsInFile(int PatCount,uint VecSize,float fbitBalance)
 		SearchTime++;
 		do
 		{
-		c = ceil((gsl_rng_uniform(mprng)+0.001)*(PatCount))-1;
+		c = ceil((double)(gsl_rng_uniform(mprng)+0.001)*(PatCount))-1;
 		}while (c==PatCount);
 
 		do
